Move violation-only helpers  to ProjectsHelper
